refactor(whispering): colocate migration files under src/lib/migration/
Migration logic was spread across state/, components/, and routes/_layout-utils/
with mixed concerns. The state/ files were plain async functions (not reactive
singletons), MigrationDialog.svelte mixed 150 lines of orchestration state with
UI, and the migrationDialog singleton was imported through the component module.

- Move migrate-database.ts and migrate-settings.ts from state/ to migration/
- Extract createMigrationDialog() into migration-dialog.svelte.ts
- Extract createMigrationTestData() into migration-test-data.ts
- Rewrite MigrationDialog.svelte as thin UI shell in migration/
- Rename check-indexeddb-migration.ts to check-database-migration.ts
- Fix pre-existing type error: properly unwrap Result from migrateDatabaseToWorkspace
- Update all 6 consumer imports
